# backend/docs/swagger.yaml
basePath: /
definitions:
...
  regions.Region:
    properties:
      capacity_mw:
        type: number
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      expected_plants:
        type: integer
      expected_users:
        type: integer
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      state:
        type: string
      status:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
  regions.RegionRequest:
    properties:
      capacity_mw:
        type: number
      country:
        type: string
      description:
        type: string
      expected_plants:
        type: integer
      expected_users:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      state:
        type: string
      status:
        type: string
      timezone:
        type: string
    required:
    - country
    - name
    - state
    - timezone
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for the Solar Energy Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Solar Energy Management System API
  version: "1.0"
paths:
...
  /superadmin/regions:
    get:
      consumes:
      - application/json
      description: Get a list of all regions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/regions.Region'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all regions
      tags:
      - Regions
    post:
      consumes:
      - application/json
      description: Create a new region
      parameters:
      - description: Region creation data
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/regions.RegionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/regions.Region'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a region
      tags:
      - Regions
  /superadmin/regions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a region by ID
      parameters:
      - description: Region ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a region
      tags:
      - Regions
    put:
      consumes:
      - application/json
      description: Update region details
      parameters:
      - description: Region ID
        in: path
        name: id
        required: true
        type: string
      - description: Region update data
        in: body
        name: region
        required: true
        schema:
          $ref: '#/definitions/regions.RegionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/regions.Region'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a region
      tags:
      - Regions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"